Use URL in AppleSampler instead of file paths
